Fix require cycle by moving push notification tracking to separate module
